Move filters to filter.py file
